error_reporting_fix

+++ core/database.php
@@ -6,7 +6,7 @@ function env($value)
       $core = parse_ini_file($env);
       return $core[$value];
 }
-error_reporting(env("ERROR_REPORT"));
+error_reporting((int) env("ERROR_REPORT"));
 $tailwind = env("TAILWIND");
 $alpineJs = env("ALPINE_JS");
 $reminderThreshold = env("REMINDER_TIME");
@@ -31,7 +31,7 @@ $db = mysqli_connect(env("HOST"), env("USER"), env("PWD"), env("DB"));
 if (isset($_SESSION['user'])):
       $userid = $_SESSION['user'];
       $userData = $db->query("SELECT * FROM `users` WHERE `id`='$userid'");
-      if (mysqli_num_rows($userData) >  0):
+      if (mysqli_num_rows($userData) > 0):
             $GetUserData = mysqli_fetch_object($userData);
             $userName = $GetUserData->username;
             $userEmail = $GetUserData->email;
